Run the tests in CI, and let the ones that should fail be seen
The suite job ran everything, which now includes the readings the engine gets
wrong, so it would have gone red for a reason no reviewer could act on. The
green gate is the suite without those, and a second job runs only those and
reports them without blocking. The count is the worklist.

The divine name is written in twelve shapes and three were tested. Each is read
on its own now, and they divide as the rule says: heavy unless a kasra in the
prefix reaches the lam.

`rasm` was a folder of five files about one thing and is one file.

The README says what a site is, what the word numbers mean, and what ibtidaa,
waqf and wasl do. The rest was commentary on decisions the layout already
shows.
